CO2 PB added
This branch contains the SolarTherm-particle branch with a CO2 power block.
The models are located in Models.PowerBlocks.sCO2Cycle and the associated media in its namesake package.
The most uptodates are the DirectDesign and QuickCO2PB.InterpolSimplePB.
The first one is implemented in the examples PhysicalParticleCO2 and the other in SimpleParticleCO2; the latter allowing a faster calculation because of model simplification.
More changes are uptocome; feedbacks from other system version would be liked, as the properties of CO2 are calculated from tables and it could be an issue.